backend/app/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
SQLAlchemy==2.0.22
sqlalchemy-orm==1.2.10
pydantic-settings==2.0.3
psycopg2==2.9.9
pydantic[email]==2.4.2
python-multipart==0.0.6
bcrypt==4.0.1
numpy==1.24.2
pandas==1.4.4
XlsxWriter==3.1.9
pyodbc==4.0.39
redis==5.0.1
QuantLib==1.32
openpyxl==3.1.2
scikit-learn==1.1.3
pyarrow==9.0.0